1.2.2

+++ upload/library/LiamW/PasswordRequirements/Extend/DataWriter/User.php
@@ -4,9 +4,7 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 {
 	public function setPassword($password, $passwordConfirm = false, XenForo_Authentication_Abstract $auth = null, $requirePassword = false)
 	{
-		if ($this->getOption(self::OPTION_ADMIN_EDIT) || ($this->isUpdate() && XenForo_Visitor::getInstance()
-					->hasPermission('general', 'lw_bypassPr'))
-		)
+		if ($this->getOption(self::OPTION_ADMIN_EDIT) || ($this->isUpdate() && XenForo_Visitor::getInstance()->hasPermission('general', 'lw_bypassPr')))
 		{
 			return parent::setPassword($password, $passwordConfirm, $auth, $requirePassword);
 		}
@@ -23,29 +21,40 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 			$errors[] = new XenForo_Phrase('liam_passwordRequirements_requested_password_blacklisted');
 		}
 
-		$auth = XenForo_Authentication_Abstract::create($this->getExisting('scheme_class'));
-		$auth->setData($this->get('data', 'xf_user_authenticate'));
+		// Check if the user is changing their password to their current password.
+		$testAuth = XenForo_Authentication_Abstract::create($this->getExisting('scheme_class'));
+		$testAuth->setData($this->get('data', 'xf_user_authenticate'));
 
-		if ($auth->authenticate($this->get('user_id'), $password))
+		if ($testAuth->authenticate($this->get('user_id'), $password))
 		{
 			$errors[] = new XenForo_Phrase('liam_passwordRequirements_you_cannot_set_your_new_password_same_as_old');
 		}
 
-		if ($historyPasswords = $passwordCriteria['password_history'])
+		// Check historic passwords
+		if ($historicPasswordCount = $passwordCriteria['password_history'])
 		{
 			$userId = $this->get('user_id');
 
-			$historicPasswords = $forcePasswordChangeModel->getHistoricPasswordDataForUser($userId,
-				$historyPasswords);
+			$historicPasswords = $forcePasswordChangeModel->getHistoricPasswordDataForUser($userId, $historicPasswordCount);
 
 			foreach ($historicPasswords AS $historicPassword)
 			{
-				$auth = XenForo_Authentication_Abstract::create($historicPassword['scheme_class']);
-				$auth->setData($historicPassword['data']);
-
-				if ($auth->authenticate($userId, $password))
+				try
 				{
-					$errors[] = new XenForo_Phrase('liam_passwordRequirements_this_password_has_been_used_before');
+					$testAuth = XenForo_Authentication_Abstract::create($historicPassword['scheme_class']);
+				} catch (XenForo_Exception $e)
+				{
+					//TODO: Check another valid old password?
+					continue;
+				}
+
+				$testAuth->setData($historicPassword['data']);
+
+				if ($testAuth->authenticate($userId, $password))
+				{
+					$errors[] = new XenForo_Phrase('liam_passwordRequirements_this_password_has_been_used_before_cannot_use_previous_x_passwords', array(
+						'count' => $historicPasswordCount
+					));
 					break;
 				}
 			}
@@ -56,8 +65,7 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 			if (($this->get('password_date') + ($passwordCriteria['min_age'] * 24 * 60 * 60)) > XenForo_Application::$time)
 			{
 				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_wait_x_days_to_change_password', array(
-					'remaining' =>
-						ceil(((($this->get('password_date') + ($passwordCriteria['min_age'] * 24 * 60 * 60)) - XenForo_Application::$time) / (24 * 60 * 60)))
+					'remaining' => ceil(((($this->get('password_date') + ($passwordCriteria['min_age'] * 24 * 60 * 60)) - XenForo_Application::$time) / (24 * 60 * 60)))
 				));
 			}
 		}
@@ -66,8 +74,7 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 		{
 			if (mb_strlen($password) < $passwordCriteria['min_length'])
 			{
-				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_be_at_least_x_characters',
-					array('count' => $passwordCriteria['min_length']));
+				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_be_at_least_x_characters', array('count' => $passwordCriteria['min_length']));
 			}
 		}
 
@@ -75,22 +82,36 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 		{
 			if (mb_strlen($password) > $passwordCriteria['min_length'])
 			{
-				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_be_less_than_x_characters',
-					array('count' => $passwordCriteria['min_length']));
+				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_be_less_than_x_characters', array('count' => $passwordCriteria['min_length']));
 			}
 		}
 
-		if ($passwordCriteria['min_special'] || $passwordCriteria['max_special'])
+		if ($passwordCriteria['min_special'] && !$passwordCriteria['max_special'])
 		{
-			if (!preg_match_all('#[^\w]#iu', $password,
-					$matches) || count($matches[0]) < $passwordCriteria['min_special'] || ($passwordCriteria['max_special'] && count($matches[0]) > $passwordCriteria['max_special'])
-			)
+			if (!preg_match_all('#[^\w]#iu', $password, $matches) || count($matches[0]) < $passwordCriteria['min_special'])
 			{
-				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_contain_between_x_and_x_special_characters',
-					array(
-						'min' => $passwordCriteria['min_special'],
-						'max' => $passwordCriteria['max_special']
-					));
+				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_contain_at_least_x_special_characters', array(
+					'count' => $passwordCriteria['min_special']
+				));
+			}
+		}
+		else if (!$passwordCriteria['min_special'] && $passwordCriteria['max_special'])
+		{
+			if (preg_match_all('#[^\w]#iu', $password, $matches) && count($matches[0]) > $passwordCriteria['max_special'])
+			{
+				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_contain_no_more_than_x_special_characters', array(
+					'count' => $passwordCriteria['max_special']
+				));
+			}
+		}
+		else if ($passwordCriteria['min_special'] && $passwordCriteria['max_special'])
+		{
+			if (!preg_match_all('#[^\w]#iu', $password, $matches) || count($matches[0]) < $passwordCriteria['min_special'] || count($matches[0]) > $passwordCriteria['max_special'])
+			{
+				$errors[] = new XenForo_Phrase('liam_passwordRequirements_must_contain_between_x_and_x_special_characters', array(
+					'min' => $passwordCriteria['min_special'],
+					'max' => $passwordCriteria['max_special']
+				));
 			}
 		}
 
@@ -110,9 +131,7 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 
 		if ($passwordCriteria['complex'])
 		{
-			if (!preg_match('/[A-Z]+/', $password) || !preg_match('/[0-9]+/', $password) || mb_stripos($password,
-					$this->get('username')) !== false || mb_strlen($password) < 8
-			)
+			if ($this->get('username') && !preg_match('/[A-Z]+/', $password) || !preg_match('/[0-9]+/', $password) || mb_stripos($password, $this->get('username')) !== false || mb_strlen($password) < 8)
 			{
 				$errors[] = new XenForo_Phrase('liam_passwordRequirements_not_meet_complexity_requirements');
 			}
@@ -120,7 +139,12 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 
 		if ($errors)
 		{
-			$this->error(reset($errors));
+			$this->error(array_shift($errors), 'password');
+
+			foreach ($errors AS $error)
+			{
+				$this->error($error);
+			}
 
 			return false;
 		}
@@ -134,8 +158,7 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 
 		if ($this->isChanged('data', 'xf_user_authenticate'))
 		{
-			$this->_db->delete('liam_pr_force_change',
-				'user_id = ' . $this->get('user_id') . ' AND initiation_date < ' . XenForo_Application::$time);
+			$this->_db->delete('liam_pr_force_change', 'user_id = ' . $this->get('user_id') . ' AND initiation_date < ' . XenForo_Application::$time);
 		}
 	}
 
@@ -148,10 +171,10 @@ class LiamW_PasswordRequirements_Extend_DataWriter_User extends XFCP_LiamW_Passw
 			$method = ($this->isInsert() ? 'getNew' : 'getExisting');
 
 			$this->_db->insert('liam_pr_password_history', array(
-				'user_id' => $this->get('user_id'),
-				'change_date' => XenForo_Application::$time,
+				'user_id'      => $this->get('user_id'),
+				'change_date'  => XenForo_Application::$time,
 				'scheme_class' => $this->$method('scheme_class'),
-				'data' => $this->$method('data', 'xf_user_authenticate')
+				'data'         => $this->$method('data', 'xf_user_authenticate')
 			));
 		}
 	}
